+++ src/sys-compote/index.js
@@ -1,0 +1,8 @@
+// Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
+// SPDX-License-Identifier: Apache-2.0
+
+export * from '#x/BaseComponent';
+export * from '#x/BaseNamedComponent';
+export * from '#x/ControlContext';
+export * from '#x/IntfComponent';
+export * from '#x/RootControlContext';