+++ src/sys-framework/export/BaseService.js
@@ -1,15 +1,14 @@
 // Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
 // SPDX-License-Identifier: Apache-2.0
 
+import { BaseNamedComponent } from '@this/sys-compote';
 import { ServiceConfig } from '@this/sys-config';
-
-import { BaseComponent } from '#x/BaseComponent';
 
 
 /**
  * Base class for system services.
  */
-export class BaseService extends BaseComponent {
+export class BaseService extends BaseNamedComponent {
   /**
    * Constructs an instance.
    *