+++ src/sys-framework/export/NetworkEndpoint.js
@@ -8,10 +8,10 @@ import { IntfRateLimiter, IntfRequestLogger, ProtocolWrangler,
   from '@this/net-protocol';
 import { DispatchInfo, IntfRequestHandler, OutgoingResponse }
   from '@this/net-util';
+import { BaseNamedComponent } from '@this/sys-compote';
 import { EndpointConfig } from '@this/sys-config';
 
 import { BaseApplication } from '#x/BaseApplication';
-import { BaseComponent } from '#x/BaseComponent';
 import { HostManager } from '#x/HostManager';
 
 
@@ -24,7 +24,7 @@ import { HostManager } from '#x/HostManager';
  *
  * @implements {IntfRequestHandler}
  */
-export class NetworkEndpoint extends BaseComponent {
+export class NetworkEndpoint extends BaseNamedComponent {
   /**
    * Application to send requests to. Becomes non-`null` during {@link
    * #_impl_start()}.