+++ src/sys-framework/export/Warehouse.js
@@ -3,17 +3,16 @@
 
 import { PromiseUtil } from '@this/async';
 import { WallClock } from '@this/clocks';
+import { BaseComponent, ControlContext, RootControlContext }
+  from '@this/sys-compote';
 import { WarehouseConfig } from '@this/sys-config';
 import { MustBe } from '@this/typey';
 
 import { BaseApplication } from '#x/BaseApplication';
-import { BaseControllable } from '#x/BaseControllable';
 import { BaseService } from '#x/BaseService';
 import { ComponentManager } from '#x/ComponentManager';
-import { ControlContext } from '#x/ControlContext';
 import { EndpointManager } from '#x/EndpointManager';
 import { HostManager } from '#x/HostManager';
-import { RootControlContext } from '#x/RootControlContext';
 import { ThisModule } from '#p/ThisModule';
 
 
@@ -25,7 +24,7 @@ import { ThisModule } from '#p/ThisModule';
  * Similarly, when `stop()`ping, the order is reversed, though the system will
  * press on with the `stop()` actions if an earlier layer is taking too long.
  */
-export class Warehouse extends BaseControllable {
+export class Warehouse extends BaseComponent {
   /**
    * Application manager.
    *