+++ CHANGELOG.md
@@ -7,7 +7,10 @@ versioning principles. Unstable releases do not.
 ### [Unreleased]
 
 Breaking changes:
-* None.
+* `sys-compote` / `sys-framework`:
+  * Extracted the lower layer of classes from `sys-framework` into new module
+    `sys-compote`. Renamed the classes for better harmony, and extracted a new
+    interface `IntfComponent`.
 
 Other notable changes:
 * None.